Use variable properties to split variable integer in bits
A variable property with the name as "bitX", with X is a positive
integer, represent the bit X of the associated variable.
Its value is a boolean.

When writing: process device write tasks with the value Y (boolean) for
a variable property with name as "bitX" of a variable A, by creating a
device write task for the variable A modifying the bit X set to Y.

When reading: save to the existing variable properties with name as
"bitX" of a variable the value of the variable read for the
corresponding bit.