add DataSourceModel, Datasource and DjangoDatabase
DataSourceModel :
Used to define a data source type.
The data source base model have a foreign key to this model to specify
the configuration :
- the name,
- can add, modify on select in the admin panel,
- the model name of the inline having the specific config (fields,
functions, manager).

DataSource :
The base model for all the data sources.
A data source needs to inherit from this class, and should have the
basic functions :
- last_value,
- read_multiple,
- write_multiple,
- get_first_element_timestamp,
- get_last_element_timestamp.

DjangoDatabase :
Specify a table to store the values. The table model should have a
manager similar to the RecordedDataManager (functions).

The default data source added is the RecordedData table.

To add new data source, look at the example in pyscada-operations.