add VariableManager, rename and update RecordedDataManager
To switch to the data source architecture, use the VariableManager
function in order to :
- filter Variable list by datasource (_get_variables_by_datasource)
- read values from datasources (read_multiple)
- write values to datasources (write_multiple)
- get first timestamp recorded for a variable list
(get_first_element_timestamp)
- get last one (get_last_element_timestamp)

Plugins and handlers should not use directly the RecordedDataManager but
the VariableManager in order to talk to differents data sources.

Look at the new way to save read data from GenericDevice and
GenericDeviceHandler in pyscada/device.py and in the pyscada.operations
plugin.