fix variable protocol save and delete
when saving a variable, the related variable protocol (modbusvariable
for example) should be saved , even if the modbus variable config is
empty.

when changing the device of a vaiable, if the protocol change, (modbus
to visa for example), the related modbus variable should be deleted and
the visa variable should be created.